add TZ to captured_date
The date_utc supplied by insights-core is TZ-naive (see
https://git.io/JvVeG). This commit adds the TZ info back onto the
datetime object before pulling the isoformat().

old format of `captured_date`: `2020-01-10T15:50:47`
new format of `captured_date`: `2020-01-10T15:50:47+00:00`

NOTE: this commit adds the `pytz` library to the Pipfile. It also
bumps the versions of other dependencies to the latest versions.

NOTE 2: this commit also adds `inventory-rest` to the
`full-stack.yml`, which listens on port 8082. This is useful for
pulling system profiles during local testing to confirm they are
landing correctly.