feat: improve puptoo parser logging (#97)
When a parser fails in system_profile we need to be able to easily
track down which one caused the problem. This new system will log each
parser when it fails and cause the payload to be rejected. We will be
able to search these failures with the request_id in kibana which will
help speed up troubleshooting time.

Signed-off-by: Stephen Adams <tsadams@gmail.com>