[feat] add check of unextracted size (#143)
* [feat] add check of unextracted size

If we get archives with more than 1GB of unextracted data, we tend to
blow up in the engine. We should not be getting that much data at a
time. This adds a check of that unextracted file size and rejects any
archives that are too large.

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* [feat] observe extraction sizes

Metrics to watch for extraction sizes over 1GB. We'll use this to
potentially reject archives in the future.

Signed-off-by: Stephen Adams <tsadams@gmail.com>