[task] add check for anemic tentants
Since puptoo is running on the new ingress `announce` topic, we need to
make sure that it doesn't accept anemic tenants until the downstream is
ready for them. The announce topic could potentially send an account
without an EBS account number and currently most downstream services
require it.

This check can be removed when the full support is available downstream.

Signed-off-by: Stephen Adams <tsadams@gmail.com>